requirements.txt
backend (pip):
django
djangorestframework
django-cors-headers

frontend (npm):
axios